requirements_cpu.txt
tensorflow-cpu
tfserve
Pillow
